+++ packages/core/src/utilities/index.ts
@@ -59,7 +59,7 @@ import { updateVTKImageDataWithCornerstoneImage } from './updateVTKImageDataWith
 import ProgressiveIterator from './ProgressiveIterator';
 import decimate from './decimate';
 import imageRetrieveMetadataProvider from './imageRetrieveMetadataProvider';
-import genericViewportDataSetMetadataProvider from './genericViewportDataSetMetadataProvider';
+import genericViewportDisplaySetMetadataProvider from './genericViewportDisplaySetMetadataProvider';
 import isVideoTransferSyntax from './isVideoTransferSyntax';
 import { getBufferConfiguration } from './getBufferConfiguration';
 import { generateVolumePropsFromImageIds } from './generateVolumePropsFromImageIds';
@@ -93,6 +93,8 @@ import { buildMetadata } from './buildMetadata';
 
 // solving the circular dependency issue
 import { _getViewportModality } from './getViewportModality';
+import { _getScalingDescriptor } from './getScalingDescriptor';
+import { resolveGenericViewportVolumeId } from './resolveGenericViewportVolumeId';
 import cache from '../cache/cache';
 import getDynamicVolumeInfo from './getDynamicVolumeInfo';
 import autoLoad from './autoLoad';
@@ -124,11 +126,18 @@ import {
   viewportSupportsVolumeCompatibility,
   viewportSupportsVolumeId,
   viewportSupportsVolumeURI,
+  isGenericViewport,
+  viewportSupportsDisplaySetPresentation,
+  getViewportContentMode,
+  viewportIsInVolumeMode,
+  viewportIsInStackMode,
 } from './viewportCapabilities';
 import { getNormalizedAspectRatio } from './getNormalizedAspectRatio';
 export { updatePlaneRestriction } from './updatePlaneRestriction';
 const getViewportModality = (viewport: IViewport, volumeId?: string) =>
   _getViewportModality(viewport, volumeId, cache.getVolume);
+const getScalingDescriptor = (viewport: IViewport, targetId?: string) =>
+  _getScalingDescriptor(viewport, targetId, cache.getVolume);
 
 export * from './isEqual';
 
@@ -153,6 +162,8 @@ export {
   getVOIRangeFromWindowLevel,
   isOpposite,
   getViewportModality,
+  getScalingDescriptor,
+  resolveGenericViewportVolumeId,
   windowLevel,
   convertToGrayscale,
   convertColorArrayToRgbString,
@@ -194,7 +205,7 @@ export {
   ProgressiveIterator,
   decimate,
   imageRetrieveMetadataProvider,
-  genericViewportDataSetMetadataProvider,
+  genericViewportDisplaySetMetadataProvider,
   transferFunctionUtils,
   updateVTKImageDataWithCornerstoneImage,
   sortImageIdsAndGetSpacing,
@@ -248,5 +259,10 @@ export {
   viewportSupportsVolumeCompatibility,
   viewportSupportsVolumeId,
   viewportSupportsVolumeURI,
+  isGenericViewport,
+  viewportSupportsDisplaySetPresentation,
+  getViewportContentMode,
+  viewportIsInVolumeMode,
+  viewportIsInStackMode,
   getNormalizedAspectRatio,
 };